Added topics to short comment and error checking to user inputs

+++ src/mapstructs/Map.java
@@ -82,7 +82,7 @@ public class Map {
 	 * @param pGamma	Inflation value for Markov Clustering
 	 * @throws IOException
 	 */
-	public void execute(File output, int pGamma) throws IOException {
+	public void execute(File output, double pGamma) throws IOException {
 		
 		this.cluBuilder(pGamma);
